+++ __tests__/plots/tooltip/flare-treemap-default.ts
@@ -18,7 +18,6 @@ export async function flareTreemapDefault(): Promise<G2Spec> {
     scale: { color: { range: schemeTableau10 } },
     encode: {
       value: 'size',
-      color: (d) => d.parent.data.name.split('.')[1],
     },
     style: {
       labelText: (d) => {